Build on msvc.

+++ components/common/tests/file_system/buffered_file1.cpp
@@ -392,7 +392,7 @@ int main ()
 
     memset (buffer,0,sizeof (buffer));
 
-    file.Read (buffer,file.Size ());
+    file.Read (buffer, (size_t)file.Size ());
 
     printf ("%s",buffer);
